Revert "Revert "split help.jl & i18n.jl out of util.jl""
This reverts commit d1a18a78db3c6d0a2326c22fafda1aee32e7775f.
